refactor visibility checker
