+++ src/query/users/src/lib.rs
@@ -27,6 +27,7 @@ mod user_mgr;
 mod user_setting;
 mod user_stage;
 mod user_udf;
+mod visibility_checker;
 
 pub mod file_format;
 pub mod idm_config;
@@ -39,3 +40,4 @@ pub use role_mgr::BUILTIN_ROLE_ACCOUNT_ADMIN;
 pub use role_mgr::BUILTIN_ROLE_PUBLIC;
 pub use user::CertifiedInfo;
 pub use user_api::UserApiProvider;
+pub use visibility_checker::GrantObjectVisibilityChecker;
